Cleanup various warnings etc from cppcheck. Remove filename path from yaiv_win, track only in XBox.

+++ src/humansize.cpp
@@ -12,14 +12,14 @@ static char *humanSize(size_t bytes, char *buf, int bufsize)
     char length = sizeof(suffix) / sizeof(suffix[0]);
 
     int i = 0;
-    double dblBytes = bytes;
+    long double dblBytes = bytes;
 
     if (bytes > 1024) {
         for (i = 0; (bytes / 1024) > 0 && i<length-1; i++, bytes /= 1024)
-            dblBytes = bytes / 1024.0;
+            dblBytes = (long double)bytes / 1024.0;
     }
 
-    snprintf(buf, bufsize, "%.02lf%s", dblBytes, suffix[i]);
+    snprintf(buf, bufsize, "%.02lf%s", (double)dblBytes, suffix[i]);
     return buf;
 }
